requirements.txt
fastapi==0.118.0
pymongo==4.13.0
motor==3.7.1
python-dotenv==1.0.1
pydantic==2.12.5
pydantic_settings==2.13.0
cryptography==42.0.5
OpenSSL==23.2.0
python-dateutil==2.9.0
